Rename files and variables for clarity
